requirements.txt
fastapi==0.115.12
pycryptodome==3.23.0
uvicorn==0.29.0
python-multipart==0.0.9